Move the resolver to the TLB
