+++ opencog/atomspace/AtomTable.cc
@@ -74,7 +74,7 @@ AtomTable::AtomTable(AtomTable* parent, AtomSpace* holder, bool transient)
     // Set resolver before doing anything else, such as getting
     // the atom-added signals.  Just in case some other thread
     // is busy adding types while we are being created.
-    Handle::set_resolver(this);
+    TLB::set_resolver(this);
 
     // Connect signal to find out about type additions
     addedTypeConnection =
@@ -87,7 +87,7 @@ AtomTable::~AtomTable()
     // Disconnect signals. Only then clear the resolver.
     std::lock_guard<std::recursive_mutex> lck(_mtx);
     addedTypeConnection.disconnect();
-    Handle::clear_resolver(this);
+    TLB::clear_resolver(this);
 
     // No one who shall look at these atoms shall ever again
     // find a reference to this atomtable.
@@ -120,7 +120,7 @@ void AtomTable::ready_transient(AtomTable* parent, AtomSpace* holder)
     _as = holder;
 
     // We can now resolve handles.
-    Handle::set_resolver(this);
+    // TLB::set_resolver(this);
 }
 
 void AtomTable::clear_transient()
@@ -130,7 +130,7 @@ void AtomTable::clear_transient()
                 "AtomTable - clear_transient called on non-transient atom table.");
 
     // We are no longer a resolver for handles.
-    Handle::clear_resolver(this);
+    // TLB::clear_resolver(this);
 
     // Clear all the atoms
     clear_all_atoms();
